requirements.txt
python-telegram-bot==12.8
pyrogram
requests
sqlalchemy
psycopg2
ujson
tgcrypto
pyfy
lyricsgenius
deezloader
mutagen
speedtest-cli>=2.0.2
